fix verifier
